Beta

+++ app/Http/Controllers/RecetaController.php
@@ -18,7 +18,7 @@ class RecetaController extends Controller
      */
     public function index()
     {
-        $recetas = Receta::all();
+        $recetas = Receta::with('categoria')->get();
         return view('recetas.recetaIndex',compact('recetas'));
     }
 
@@ -52,7 +52,7 @@ class RecetaController extends Controller
 
         //dd($request->all());
         $receta = Receta::create($request->all());
-        return redirect()->route('recetas.ingrediente',[$receta->id]);
+        return redirect()->route('recetas.ingredientes',[$receta->id]);
     }
 
     /**
@@ -74,6 +74,7 @@ class RecetaController extends Controller
      */
     public function edit(Receta $receta)
     {
+        $this->authorize('update',$receta);
         $categorias = Categoria::all();
         return view('recetas.recetaForm',compact('receta','categorias'));
     }
@@ -87,6 +88,7 @@ class RecetaController extends Controller
      */
     public function update(Request $request, Receta $receta)
     {
+        $this->authorize('update',$receta);
         $request->validate([
             'titulo' => ['string','max:255','required'],
             'descripcion' => ['required'],
@@ -97,7 +99,7 @@ class RecetaController extends Controller
         ]);
 
         Receta::where('id',$receta->id)->update($request->except('_token','_method'));
-        return redirect()->route('recetas.ingrediente',[$receta->id]);
+        return redirect()->route('recetas.ingredientes',[$receta->id]);
     }
 
     /**
@@ -108,14 +110,14 @@ class RecetaController extends Controller
      */
     public function destroy(Receta $receta)
     {
+        $this->authorize('delete',$receta);
         $receta->delete();
         return redirect()->route('recetas.index');
 
     }
 
-    public function agregarProcesos($id)
+    public function agregarProcesos(Receta $receta)
     {
-        $receta = Receta::find($id);
         return view("recetas.recetaAgregarProcesos",compact('receta'));
     }
 
@@ -125,16 +127,26 @@ class RecetaController extends Controller
         return view("recetas.recetaAgregarIngredientes",compact('receta'));
     }
 
-    public function agregarIngrediente(Request $request, Receta $receta)
+    public function agregarIngrediente(Request $request,Receta $receta)
     {
+        $request->validate([
+            'nombre' => ['string','max:255','required'],
+            'cantidad' => ['string','max:255','required'],
+        ]);
         $ingrediente = Ingrediente::firstOrCreate(['nombre' => $request->nombre]);
-        $receta->ingredientes()->attach($ingrediente->id);
-        return redirect()->route('recetas.ingrediente',[$receta->id]);
+        $receta->ingredientes()->attach($ingrediente->id,['cantidad' => $request->cantidad]);
+        return redirect()->route('recetas.ingredientes',[$receta->id])->with([
+            'mensaje' => 'Ingrediente aregado Correctamente',
+            'alert-type' => 'alert-info',
+        ]);
     }
 
     public function eliminarIngrediente(Request $request, Receta $receta)
     {
-        $receta->ingredientes()->attach($request->only('id'));
-        return redirect()->route('recetas.ingrediente',[$receta->id]);
+        $receta->ingredientes()->detach($request->only('id'));
+        return redirect()->route('recetas.ingredientes',[$receta->id])->with([
+            'mensaje' => 'Ingrediente eliminado de la receta',
+            'alert-type' => 'alert-info',
+        ]);;
     }
 }
